fix name LogoutBtn
